+++ src/main/java/de/stamme/basicquests/commands/SkipQuestCommand.java
@@ -144,7 +144,7 @@ public class SkipQuestCommand implements CommandExecutor {
 		return true;
 	}
 
-	/** // TODO
+	/**
 	 * Called when a CommandSender tries to skip a QuestPlayers quest by it's index
 	 * sender and target can be the same player!
 	 * sender -> /skipquest <target> [questIndex]