+++ src/main/java/de/stamme/basicquests/model/generation/QuestGenerationException.java
@@ -4,7 +4,7 @@ public class QuestGenerationException extends Exception {
 
 	private static final long serialVersionUID = -2941555618699589814L;
 
-	public String message;
+	public final String message;
 	
 	public QuestGenerationException(String message) {
 		this.message = message;