+++ src/main/java/de/stamme/basicquests/model/quests/ReachLevelQuest.java
@@ -1,7 +1,6 @@
 package de.stamme.basicquests.model.quests;
 
 import de.stamme.basicquests.model.QuestPlayer;
-import de.stamme.basicquests.util.StringFormatter;
 
 public class ReachLevelQuest extends Quest {