+++ src/main/java/de/stamme/basicquests/model/wrapper/structure/QuestStructureService_1_18.java
@@ -8,8 +8,8 @@ import org.jetbrains.annotations.Nullable;
 import java.util.Collections;
 import java.util.List;
 
-// Only compatible with spigot 1.16 - 1.18!
-public class QuestStructureService_1_16 extends QuestStructureService {
+// In 1.18 only VILLAGE, RUINED_PORTAL, SHIPWRECK, and OCEAN RUINS can be located...
+public class QuestStructureService_1_18 extends QuestStructureService {
 
     @Override
     public @Nullable Location findStructureNearLocation(QuestStructureType questStructureType, Location nearLocation, World world) {
@@ -22,6 +22,7 @@ public class QuestStructureService_1_16 extends QuestStructureService {
             if (structureType == null) continue;
 
             Location location = world.locateNearestStructure(nearLocation, structureType, 100, false);
+
             if (location == null) continue;
 
             double distance = Math.sqrt(Math.abs(nearLocation.getX() - location.getX()) + Math.abs(nearLocation.getZ() - location.getZ()));
@@ -30,6 +31,8 @@ public class QuestStructureService_1_16 extends QuestStructureService {
                 nearestLocation = location;
             }
         }
+
+        assert nearestLocation != null;
 
         return nearestLocation;
     }
@@ -71,6 +74,46 @@ public class QuestStructureService_1_16 extends QuestStructureService {
                 return Collections.singletonList(StructureType.RUINED_PORTAL);
             case BASTION_REMNANT:
                 return Collections.singletonList(StructureType.BASTION_REMNANT);
+            default:
+                return null;
+        }
+    }
+
+    /**
+     * In 1.18 only a handful of structures can be found. Therefore only quests with these structures can be generated.
+     */
+    @Override
+    @Nullable
+    public QuestStructureType fromString(String structureName) {
+        switch (structureName.toUpperCase()) {
+            case "VILLAGE":
+            case "VILLAGE_PLAINS":
+            case "VILLAGE_DESERT":
+            case "VILLAGE_SAVANNA":
+            case "VILLAGE_SNOWY":
+            case "VILLAGE_TAIGA":
+                // Not differentiating between different villages.
+                return QuestStructureType.VILLAGE;
+
+            case "OCEAN_RUIN": // bukkit 1.16
+            case "OCEAN_RUIN_WARM":
+            case "OCEAN_RUIN_COLD":
+                // Not differentiating between different ocean ruins.
+                return QuestStructureType.OCEAN_RUIN;
+
+            case "SHIPWRECK":
+            case "SHIPWRECK_BEACHED":
+                return QuestStructureType.SHIPWRECK;
+
+            case "RUINED_PORTAL":
+            case "RUINED_PORTAL_DESERT":
+            case "RUINED_PORTAL_JUNGLE":
+            case "RUINED_PORTAL_MOUNTAIN":
+            case "RUINED_PORTAL_NETHER":
+            case "RUINED_PORTAL_OCEAN":
+            case "RUINED_PORTAL_SWAMP":
+                // Not differentiating between different ruined portals.
+                return QuestStructureType.RUINED_PORTAL;
             default:
                 return null;
         }